Move pages schema to project-build
We are moving all build related code from project to project-build.
Here moved pages schema and Build type.
