+++ apps/builder/app/builder/features/sidebar-left/panels/pages/pages.tsx
@@ -19,7 +19,7 @@ import {
   NewPageIcon,
   PageIcon,
 } from "@webstudio-is/icons";
-import type { Page, Pages } from "@webstudio-is/project";
+import type { Page, Pages } from "@webstudio-is/project-build";
 import type { Publish } from "~/shared/pubsub";
 import {
   useCurrentPageId,
