+++ packages/project-build/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./schema/pages";
 export * from "./schema/instances";
 export * from "./schema/props";
 export * from "./schema/styles";
