+++ packages/project/src/db/tree.ts
@@ -7,10 +7,9 @@ import {
   Instance,
   Instances,
 } from "@webstudio-is/project-build";
-import { prisma, type Prisma } from "@webstudio-is/prisma-client";
+import { type Project, type Prisma, prisma } from "@webstudio-is/prisma-client";
 import { utils } from "../index";
 import { parseProps, serializeProps } from "./props";
-import type { Project } from "../shared/schema";
 import {
   authorizeProject,
   type AppContext,
